Can't overload 64-bit alias cast, but fix for Mac

+++ include/statevector.hpp
@@ -142,12 +142,15 @@ public:
 
     complex read(const bitCapInt& i) { return read((size_t)i); }
 #if ENABLE_COMPLEX_X2
-    complex2 read2(const bitCapInt& i1, const bitCapInt& i2) { return read2((size_t)i1, (size_t)i2); }
+    complex2 read2(const bitCapInt& i1, const bitCapInt& i2)
+    {
+        return read2((size_t)(uint64_t)i1, (size_t)(uint64_t)i2);
+    }
 #endif
-    void write(const bitCapInt& i, const complex& c) { write((size_t)i, c); }
+    void write(const bitCapInt& i, const complex& c) { write((size_t)(uint64_t)i, c); }
     void write2(const bitCapInt& i1, const complex& c1, const bitCapInt& i2, const complex& c2)
     {
-        write2((size_t)i1, c1, (size_t)i2, c2);
+        write2((size_t)(uint64_t)i1, c1, (size_t)(uint64_t)i2, c2);
     }
 
     complex read(const size_t& i) { return amplitudes.get()[i]; };
